+++ common/Animations/KeyframedValueExtensions.cs
@@ -12,11 +12,11 @@ public static class KeyframedValueExtensions
     /// <summary>
     /// Performs <paramref name="action"/> for each value in <paramref name="keyframes"/> that is true.
     /// </summary>
-    public static void ForEachFlag(this KeyframedValue<bool> keyframes, Action<double, double> action)
+    public static void ForEachFlag(this KeyframedValue<bool> keyframes, Action<float, float> action)
     {
         var active = false;
-        var startTime = 0d;
-        var lastKeyframeTime = 0d;
+        var startTime = 0f;
+        var lastKeyframeTime = 0f;
 
         foreach (var keyframe in keyframes) if (keyframe.Value != active)
             {
@@ -42,7 +42,7 @@ public static class KeyframedValueExtensions
     ///<param name="x"> The <see cref="CommandScale.X"/> value of the keyframe. </param>
     ///<param name="y"> The <see cref="CommandScale.Y"/> value of the keyframe. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<CommandScale> Add(this KeyframedValue<CommandScale> keyframes, double time, double x, double y, Func<double, double> easing = null)
+    public static KeyframedValue<CommandScale> Add(this KeyframedValue<CommandScale> keyframes, float time, float x, float y, Func<float, float> easing = null)
         => keyframes.Add(time, new(x, y), easing);
 
     ///<summary> Adds a <see cref="Vector2"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -50,7 +50,7 @@ public static class KeyframedValueExtensions
     ///<param name="time"> The time of the <see cref="Keyframe{T}"/>. </param>
     ///<param name="scale"> The scale value of this <see cref="Keyframe{T}"/>. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<CommandScale> Add(this KeyframedValue<CommandScale> keyframes, double time, double scale, Func<double, double> easing = null)
+    public static KeyframedValue<CommandScale> Add(this KeyframedValue<CommandScale> keyframes, float time, float scale, Func<float, float> easing = null)
         => keyframes.Add(time, new(scale), easing);
 
     ///<summary> Adds a <see cref="CommandPosition"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -59,7 +59,7 @@ public static class KeyframedValueExtensions
     ///<param name="x"> The <see cref="CommandPosition.X"/> value of the keyframe. </param>
     ///<param name="y"> The <see cref="CommandPosition.Y"/> value of the keyframe. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<CommandPosition> Add(this KeyframedValue<CommandPosition> keyframes, double time, double x, double y, Func<double, double> easing = null)
+    public static KeyframedValue<CommandPosition> Add(this KeyframedValue<CommandPosition> keyframes, float time, float x, float y, Func<float, float> easing = null)
         => keyframes.Add(time, new(x, y), easing);
 
     ///<summary> Adds a <see cref="CommandPosition"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -67,7 +67,7 @@ public static class KeyframedValueExtensions
     ///<param name="time"> The time of the <see cref="Keyframe{T}"/>. </param>
     ///<param name="xy"> The x and y value of this <see cref="Keyframe{T}"/>. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<CommandPosition> Add(this KeyframedValue<CommandPosition> keyframes, double time, double xy, Func<double, double> easing = null)
+    public static KeyframedValue<CommandPosition> Add(this KeyframedValue<CommandPosition> keyframes, float time, float xy, Func<float, float> easing = null)
         => keyframes.Add(time, new(xy, xy), easing);
 
     ///<summary> Adds a <see cref="Vector2"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -76,7 +76,7 @@ public static class KeyframedValueExtensions
     ///<param name="x"> The <see cref="Vector2.X"/> value of the keyframe. </param>
     ///<param name="y"> The <see cref="Vector2.Y"/> value of the keyframe. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Vector2> Add(this KeyframedValue<Vector2> keyframes, double time, float x, float y, Func<double, double> easing = null)
+    public static KeyframedValue<Vector2> Add(this KeyframedValue<Vector2> keyframes, float time, float x, float y, Func<float, float> easing = null)
         => keyframes.Add(time, new(x, y), easing);
 
     ///<summary> Adds a <see cref="Vector2"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -84,7 +84,7 @@ public static class KeyframedValueExtensions
     ///<param name="time"> The time of the <see cref="Keyframe{T}"/>. </param>
     ///<param name="scale"> The scale value of this <see cref="Keyframe{T}"/>. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Vector2> Add(this KeyframedValue<Vector2> keyframes, double time, float scale, Func<double, double> easing = null)
+    public static KeyframedValue<Vector2> Add(this KeyframedValue<Vector2> keyframes, float time, float scale, Func<float, float> easing = null)
         => keyframes.Add(time, new(scale), easing);
 
     ///<summary> Adds a <see cref="Vector3"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -94,7 +94,7 @@ public static class KeyframedValueExtensions
     ///<param name="y"> The <see cref="Vector3.Y"/> value of the keyframe. </param>
     ///<param name="z"> The <see cref="Vector3.Z"/> value of the keyframe. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Vector3> Add(this KeyframedValue<Vector3> keyframes, double time, float x, float y, float z, Func<double, double> easing = null)
+    public static KeyframedValue<Vector3> Add(this KeyframedValue<Vector3> keyframes, float time, float x, float y, float z, Func<float, float> easing = null)
         => keyframes.Add(time, new(x, y, z), easing);
 
     ///<summary> Adds a <see cref="Vector3"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -102,7 +102,7 @@ public static class KeyframedValueExtensions
     ///<param name="time"> The time of the <see cref="Keyframe{T}"/>. </param>
     ///<param name="scale"> The scale value of this <see cref="Keyframe{T}"/>. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Vector3> Add(this KeyframedValue<Vector3> keyframes, double time, float scale, Func<double, double> easing = null)
+    public static KeyframedValue<Vector3> Add(this KeyframedValue<Vector3> keyframes, float time, float scale, Func<float, float> easing = null)
         => keyframes.Add(time, new(scale), easing);
 
     ///<summary> Adds a <see cref="Quaternion"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -111,9 +111,9 @@ public static class KeyframedValueExtensions
     ///<param name="axis"> The axis to rotate about. </param>
     ///<param name="angle"> The rotation angle in radians. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Quaternion> Add(this KeyframedValue<Quaternion> keyframes, double time, Vector3 axis, double angle, Func<double, double> easing = null)
+    public static KeyframedValue<Quaternion> Add(this KeyframedValue<Quaternion> keyframes, float time, Vector3 axis, float angle, Func<float, float> easing = null)
     {
-        var half = (float)angle * .5f;
+        var half = angle * .5f;
         var sin = MathF.Sin(half);
         return keyframes.Add(time, new(axis.X * sin, axis.Y * sin, axis.Z * sin, MathF.Cos(half)), easing);
     }
@@ -123,7 +123,7 @@ public static class KeyframedValueExtensions
     ///<param name="time"> The time of the <see cref="Keyframe{T}"/>. </param>
     ///<param name="angle"> The rotation angle in radians (rotates about all axes). </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Quaternion> Add(this KeyframedValue<Quaternion> keyframes, double time, float angle, Func<double, double> easing = null)
+    public static KeyframedValue<Quaternion> Add(this KeyframedValue<Quaternion> keyframes, float time, float angle, Func<float, float> easing = null)
         => keyframes.Add(time, Quaternion.CreateFromYawPitchRoll(angle, angle, angle), easing);
 
     ///<summary> Adds a <see cref="Quaternion"/> keyframe to <paramref name="keyframes"/>. </summary>
@@ -133,6 +133,6 @@ public static class KeyframedValueExtensions
     ///<param name="yaw"> The yaw (y-axis) angle of the <see cref="Quaternion"/>. </param>
     ///<param name="roll"> The roll (z-axis) angle of the <see cref="Quaternion"/>. </param>
     ///<param name="easing"> The <see cref="EasingFunctions"/> to apply to this <see cref="Keyframe{T}"/>. </param>
-    public static KeyframedValue<Quaternion> Add(this KeyframedValue<Quaternion> keyframes, double time, float pitch, float yaw, float roll, Func<double, double> easing = null)
+    public static KeyframedValue<Quaternion> Add(this KeyframedValue<Quaternion> keyframes, float time, float pitch, float yaw, float roll, Func<float, float> easing = null)
         => keyframes.Add(time, Quaternion.CreateFromYawPitchRoll(pitch, yaw, roll), easing);
 }
